more blind outlining
