--- src/main/java/com/tommy/securenote/Constants/Constants.java
@@ -1,13 +1,0 @@
-package com.tommy.securenote.Constants;
-
-public class Constants {
-
-    private Constants() {
-        // restrict instansiation
-    }
-
-    // URL MAPPINGS
-    public static final String INDEX = "/";
-    public static final String FETCH_NOTE = "/fetch";
-    public static final String CREATE_NOTE = "/create";
-}
