+++ src/main/java/com/tommy/securenote/config/MvcConfig.java
@@ -1,0 +1,15 @@
+package com.tommy.securenote.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+
+/**
+ *  Not sure what I'm doing here yet.
+ *  https://spring.io/guides/gs/securing-web/
+ */
+
+
+public class MvcConfig {
+}
