+++ src/main/java/com/tommy/securenote/controller/JsonController.java
@@ -1,0 +1,9 @@
+package com.tommy.securenote.controller;
+
+/**
+ *  Splitting out controllers. I think we should have HTML/JSON controllers with mirrored request mappings
+ *  so that a browser or application could consume the app.
+ */
+
+public class JsonController {
+}
