+++ src/main/java/com/tommy/securenote/service/RetrieveSecureNoteService.java
@@ -1,0 +1,6 @@
+package com.tommy.securenote.service;
+import org.springframework.stereotype.Service;
+
+@Service
+public class RetrieveSecureNoteService {
+}
