+++ src/main/resources/templates/index.html
@@ -1,0 +1,4 @@
+<!DOCTYPE HTML>
+<html>
+
+</html>
